like

+++ juejin-pin/src/main/java/com/bytedance/juejin/pin/domain/comment/CommentService.java
@@ -1,6 +1,7 @@
 package com.bytedance.juejin.pin.domain.comment;
 
 import com.bytedance.juejin.basic.domain.DomainEventPublisher;
+import com.bytedance.juejin.basic.exception.JuejinNotFoundException;
 import com.bytedance.juejin.pin.domain.comment.event.CommentCreatedEvent;
 import com.bytedance.juejin.pin.domain.comment.event.CommentDeletedEvent;
 import com.bytedance.juejin.pin.domain.comment.view.CommentCreateCommand;
@@ -29,6 +30,9 @@ public class CommentService {
 
     public void delete(CommentDeleteCommand delete, User user) {
         Comment comment = commentRepository.get(delete.getId());
+        if (comment == null) {
+            throw new JuejinNotFoundException(Comment.class, delete.getId());
+        }
         commentRepository.delete(comment);
         eventPublisher.publish(new CommentDeletedEvent(comment, user));
     }
